Contact manipulation .php files now handle errors
The addContact.php and searchContacts.php files now check if the UserID is in use, and the deleteContact.php and modifyContact.php files now check if the UserID and ID fields match an entry in the contact database. In both instances, while bad requests would have failed regardless, they now respond with an actual error.

+++ deleteContact.php
@@ -1,8 +1,4 @@
 <?php
-	// Enable detailed error reporting for debugging
-	ini_set('display_errors', 1);
-	ini_set('display_startup_errors', 1);
-	error_reporting(E_ALL);
 	$inData = getRequestInfo();
 	
 	$ID = $inData["ID"];
@@ -15,7 +11,22 @@
 	} 
 	else
 	{
-		//Delete the contact only if the user actually owns it.
+		// Check if the user/contact combination actually exists.
+		$stmt = $conn->prepare("select count(*) as count from Contacts where ID=? AND UserID=?");
+		$stmt->bind_param("ii", $ID, $userID);
+		if (!$stmt->execute())
+            returnWithError($stmt->error);
+
+		$result = $stmt->get_result();
+		
+		if ($result->fetch_assoc()['count'] != 1) {
+			returnWithError("User and contact pair not found.");
+			$stmt->close();
+			$conn->close();
+			return;
+		}
+
+		// Delete the contact only if the user actually owns it.
         $stmt = $conn->prepare("delete from Contacts where ID=? AND UserID=?");
         $stmt->bind_param("ii", $ID, $userID);
 		if (!$stmt->execute())
